Restructure app folders
