+++ src/components/cryptos/Currency.js
@@ -1,5 +1,5 @@
 import { NavLink, Link } from 'react-router-dom';
-import logo from '../assets/crypto.png';
+import logo from '../../assets/crypto.png';
 import { FiArrowRightCircle } from "react-icons/fi";
 
 const Currency = () => {
